+++ src/CPPMoMinterface/content/CPPMoMsolver/src/triangle.h
@@ -7,7 +7,7 @@
 class Triangle
 {
     public:
-        Triangle(int vertex_1, int vertex_2, int vertex_3, Node centre, float area);
+        Triangle(int vertex_1, int vertex_2, int vertex_3, Node centre, double area);
         Triangle(int vertex_1, int vertex_2, int vertex_3,
                     Node node_vertex_1, Node node_vertex_2, Node node_vertex_3);
 
@@ -17,7 +17,7 @@ class Triangle
         std::vector<int> getVertices();
         std::vector<int> getEdges();
 
-        float getArea();
+        double getArea();
         Node getCentre();
 
         void setEdgeIndex(int index);
@@ -26,11 +26,11 @@ class Triangle
         int vertex_1;
         int vertex_2;
         int vertex_3;
-        float area;
+        double area;
         Node centre;
         std::vector<int> edge_indices;
 
-        float calculateArea(Node node_vertex_1, Node node_vertex_2, Node node_vertex_3);
+        double calculateArea(Node node_vertex_1, Node node_vertex_2, Node node_vertex_3);
         Node calculateCentre(Node node_vertex_1, Node node_vertex_2, Node node_vertex_3);
 };
